requirements.txt
ansible==6.3.0
ansible-core==2.13.3
appdirs==1.4.4
arrow==1.2.2
asgiref==3.5.2
astroid==2.12.5
attrs==22.1.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bcrypt==4.0.0
beautifulsoup4==4.11.1
black==22.8.0
blessed==1.19.1
botocore==1.27.66
bs4==0.0.1
cached-property==1.5.2
cement==3.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.3
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.6
distro==1.7.0
Django==4.1.2
django-extensions==3.2.0
django-jquery==3.1.0
django-phonenumber-field==6.4.0
django-ses==3.3.0
djangorestframework==3.13.1
docker==6.0.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
feedgen==0.9.0
filelock==3.8.0
freezegun==1.2.2
future==0.18.2
geographiclib==1.49
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipython-genutils==0.2.0
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.15.0
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
lxml==4.9.1
markdown2==2.4.3
MarkupSafe==2.1.1
mccabe==0.7.0
more-itertools==8.14.0
mypy-extensions==0.4.3
numpy==1.23.2
packaging==21.3
paramiko==2.11.0
pathspec==0.10.1
pexpect==4.8.0
phonenumbers==8.12.54
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.3
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pylint==2.15.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytoolconfig==1.2.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
resolvelib==0.8.1
rope==1.3.0
semantic-version==2.10.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
traitlets==5.3.0
twilio==7.13.0
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.4
wcwidth==0.2.5
websocket-client==1.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
